+++ src/services-hooks/reports/useGetAllOwnersReport.tsx
@@ -63,29 +63,158 @@ export default function useGetAllOwnersReport({
         setPagination(foundPage?.pagination_data);
         dispatch(updateOwnersReportReport({ data: foundPage?.data }));
       } else {
-        const response = await axios.get(`/admin/report/owners?${queryString}`);
-        const { owners_report } = response?.data?.data;
-        const { data, current_page, last_page, per_page, total, from, to } =
-          owners_report;
-        const paginationDataset = {
-          current_page,
-          last_page,
-          per_page,
-          total,
-          from,
-          to,
-          length: data?.length,
+        // const response = await axios.get(`/admin/report/owners?${queryString}`);
+        // const { owners_report } = response?.data?.data;
+        // const { data, current_page, last_page, per_page, total, from, to } =
+        //   owners_report;
+        // const paginationDataset = {
+        //   current_page,
+        //   last_page,
+        //   per_page,
+        //   total,
+        //   from,
+        //   to,
+        //   length: data?.length,
+        // };
+        const dataset = [
+          {
+            id: 1,
+            building_id: 1,
+            building: {
+              id: 1,
+              name: "Lekki",
+              is_deleted: 0,
+              deleted_at: "",
+              deleted_by: "",
+              created_at: "",
+              updated_at: "",
+            },
+            shortlet_id: 1,
+            shortlet: {
+              id: 1,
+              name: "Top Apartment",
+              slug: "",
+              description: "",
+              location: "",
+              currency: "",
+              price: 500,
+              caution_fee: 200,
+              tax_fee: 10,
+              no_of_bedrooms: 2,
+              no_of_bathrooms: 2,
+              min_guests: "",
+              max_guests: 2,
+              point_of_interest: "",
+              cancellation_policy: "",
+              availability_status: "",
+              created_at: "",
+              updated_at: "",
+              room_option_id: 2,
+              bookings_count: 2,
+              last_booking_date: "",
+              no_of_bookings: 2,
+            },
+            expense_id: 1,
+            expense: {
+              id: 1,
+              name: "Lekki",
+              slug: "",
+              deleted_at: "",
+              deleted_by: "",
+              created_at: "",
+              updated_at: "",
+            },
+            amount: 200,
+            date: "",
+            additional_note: "",
+            monthly_amount: {
+              jan: 1,
+              feb: 1,
+              mar: 1,
+              apr: 1,
+              may: 1,
+              jun: 1,
+              jul: 1,
+              aug: 1,
+              sep: 1,
+              oct: 1,
+              nov: 1,
+              dec: 1,
+            },
+          },
+        ];
+        const summaries = {
+          monthly_totals: {
+            jan: 1,
+            feb: 1,
+            mar: 1,
+            apr: 1,
+            may: 1,
+            jun: 1,
+            jul: 1,
+            aug: 1,
+            sep: 1,
+            oct: 1,
+            nov: 1,
+            dec: 1,
+          },
+          monthly_revenue: {
+            jan: 1,
+            feb: 1,
+            mar: 1,
+            apr: 1,
+            may: 1,
+            jun: 1,
+            jul: 1,
+            aug: 1,
+            sep: 1,
+            oct: 1,
+            nov: 1,
+            dec: 1,
+          },
+          monthly_management_fee: {
+            jan: 1,
+            feb: 1,
+            mar: 1,
+            apr: 1,
+            may: 1,
+            jun: 1,
+            jul: 1,
+            aug: 1,
+            sep: 1,
+            oct: 1,
+            nov: 1,
+            dec: 1,
+          },
+          monthly_profit: {
+            jan: 1,
+            feb: 1,
+            mar: 1,
+            apr: 1,
+            may: 1,
+            jun: 1,
+            jul: 1,
+            aug: 1,
+            sep: 1,
+            oct: 1,
+            nov: 1,
+            dec: 1,
+          },
         };
-        dispatch(updateOwnersReportReport({ data }));
-        if (!remakeRequest) {
-          dispatch(
-            addToPaginationHistory({
-              pagination_data: paginationDataset,
-              data: data,
-            })
-          );
-        }
-        setPagination(paginationDataset);
+        dispatch(
+          updateOwnersReportReport({
+            data: { data: dataset, summary: summaries },
+          })
+        );
+        // if (!remakeRequest) {
+        //   dispatch(
+        //     addToPaginationHistory({
+        //       pagination_data: paginationDataset,
+        //       data: data,
+        //     })
+        //   );
+        // }
+        // setPagination(paginationDataset);
       }
     } catch (error) {
       setIsFailed(true);